Move stream struct to stream.go
