+++ server/partition.go
@@ -59,24 +59,6 @@ func (r *replica) getLatestOffset() int64 {
 	return r.offset
 }
 
-// stream is a message stream consisting of one or more partitions. Each
-// partition maps to a NATS subject and is the unit of replication.
-type stream struct {
-	name       string
-	subject    string
-	partitions map[int32]*partition
-}
-
-// Close the stream by closing each of its partitions.
-func (p *stream) Close() error {
-	for _, partition := range p.partitions {
-		if err := partition.Close(); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 // partition represents a replicated message stream partition backed by a
 // durable commit log. A partition is attached to a NATS subject and stores
 // messages on that subject in a file-backed log. A partition has a set of
