+++ server/stream.go
@@ -1,0 +1,19 @@
+package server
+
+// stream is a message stream consisting of one or more partitions. Each
+// partition maps to a NATS subject and is the unit of replication.
+type stream struct {
+	name       string
+	subject    string
+	partitions map[int32]*partition
+}
+
+// Close the stream by closing each of its partitions.
+func (p *stream) Close() error {
+	for _, partition := range p.partitions {
+		if err := partition.Close(); err != nil {
+			return err
+		}
+	}
+	return nil
+}
